FIX: tweak to upgrade script for field/resource type decoupling code [q12841]
git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20220704_dhuby_fields_belong_to_multiple_types_q12841@22099 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ upgrade/scripts/025_populate_resource_type_field_resource_type.php
@@ -1,6 +1,22 @@
 <?php
 set_sysvar(SYSVAR_UPGRADE_PROGRESS_SCRIPT, "Start decoupling of resource types and fields to enable multi type support...");
 
+$field_tinfo = ps_query('DESCRIBE resource_type_field');
+if(!is_array($field_tinfo))
+    {
+    $msg = '[error] Unable to describe table "resource_type_field"';
+    logScript($msg);
+    message_add($notification_users, "Upgrade script 025: {$msg}", '', null, MESSAGE_ENUM_NOTIFICATION_TYPE_SCREEN, MESSAGE_DEFAULT_TTL_SECONDS);
+    return;
+    }
+
+if(!in_array("resource_type",array_column($field_tinfo,"Field")))
+    {
+    // No resource_type column, script not required
+    set_sysvar(SYSVAR_UPGRADE_PROGRESS_SCRIPT, "Finished decoupling of resource types and fields to enable multi type support!");
+    return;
+    }
+
 // Clear out any entries added by dbstruct
 ps_query("TRUNCATE resource_type_field_resource_type");
